Add scoring based on keeping tiles on the edge of the board

+++ src/main/scala/TwentyFortyEight/scorers/BoardScorer.scala
@@ -3,5 +3,5 @@ package TwentyFortyEight.scorers
 import TwentyFortyEight.GameBoard
 
 trait BoardScorer {
-  def scoreSingleGameBoard(gameBoard: GameBoard): Int
+  def scoreSingleGameBoard(gameBoard: GameBoard): Double
 }
